requirements.txt
# Combined dependency requirements for the project
numpy>=1.25.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
h5py>=3.9.0
hdbscan>=0.8.29
umap-learn>=0.5.3
torch>=2.0.0
pyyaml>=6.0.0
seaborn>=0.12.0